+++ python/CHANGELOG.md
@@ -14,6 +14,7 @@ We structure this changelog in accordance with [Keep a Changelog](https://keepac
 
 [^1]: Concurrent bucket-to-bucket copy and transformation jobs is a new functionality in AIS that has been pushed upstream but not yet officially released. Please use a cluster built with the latest changes (see [here](https://github.com/NVIDIA/aistore/blob/main/docs/getting_started.md#from-source)).
 
+- Add `direct put` support in the ETL fastapi web server.
 - Add support for extension re-mapping in bucket-to-bucket copy jobs using the `ext` parameter.
 
 ### Changed